requirements_minimal.txt
# CESAROPS Minimal Requirements - Essential components only
# This file contains only the absolutely necessary dependencies

# Core data processing
pandas>=1.5.0
numpy>=1.21.0

# HTTP requests and web data
requests>=2.28.0

# Configuration files
PyYAML>=6.0

# Scientific computing
scipy>=1.9.0

# GUI framework - usually included with Python
# tkinter is built-in, no need to install

# Optional but recommended:
# simplekml>=1.3.5          # For KML export (uncomment if needed)
# scikit-learn>=1.1.0       # For ML features (uncomment if needed)  
# joblib>=1.2.0             # For ML model persistence (uncomment if needed)

# Do NOT install these if having issues:
# matplotlib  # Can cause installation problems
# plotly      # Not essential for core functionality
# jupyter     # Not needed for this application
# notebook    # Not needed for this application
